First area + Shortcut from the key + Hidden room
Added the first area after the starting area. On the east side of the first area the player can come from there as a shortcut after receiving the key of the level, instead of running all the way back (still needs work). The first area has a hidden room that appears when the player uses the ability. Hidden room could contain items or other stuff.