Area with a pit added
This area will have a death pit, the player will have to choose one of two paths. One is easier but it will require them to use their ability that will cause to bring enemies to them, the other is jumping between platforms and doesn't need the use of the ability.