Added more to the level
Added an the player will fight some enemies in there if he chooses that path. I will add an invisble bridge on top of the area, the player will have to use his ability to make appear, using the bridge he can bypass the enemies. That way the player will be able to choose how to cross the area.