Added to my level
Added a spire that will contain the key item needed. There is a path leading to it, it will have few walls that needs the ability to bypass them. Before the walls there will be a different path the player can take if they don't want to use the ability, one will lead to a platforming section and the other to a room with static enemies.